remove perl dependencies for analysis
In hindsight keeping the analysis stuff as part of the dbt2 kit or a
less ambitious package, not dbttools, just for the dbt test kits would
have been easier to maintain.  The fact that the other aspects of the
perl modules being neglected was also a factor.  The new scripts are an
adaptation of the original bash/perl/gnuplot scripts that now use use
python and R.