use rpy
Use the Python R module to do the post processing calculations.  This
does required rpy to be installed onto the driver system.  When
incporating the rpy changes to the post processing script, discovered
some dead code and that the transaction calculations were not on the
steady state duration.  Calculations were being incorrectly done on the
entire transaction log.