conform drizzle scripts
Make the usage of the drizzle scripts match how the other database
scripts work a little more closely and update the README-DRIZZLE.

Use the environment DATADIR to set the dbt2 database location in order
for the start and stop scripts to work correctly.  Open to other ways of
doing this without hard coding it in the scripts.

The only database parameters that are hardcoded are --user and
--basedir.  Everything else can be set by using the -p flag.
dbt2-drizzle-admin, dbt2-drizzle-start-db and dbt2-drizzle-stop-db have
been updated appropriately.

Use the Drizzle datadir for the pid file and database log.  This helps
accommodate installations where the user doesn't have write access to
the drizzle binary locations.

Added a flag to the dbt2-drizzle-admin script to set the database port
for client use in conjunction with the database parameter that set the
listener address, if so desired.