Simplify report script
Save the RDBMS and database name used in the readme.txt file so that the
report script can get information from the test data as opposed to
having the user remember what was used in the case of manually executing
these scripts.

Also fixed up the report script to redirect R warnings and errors to a
log file but to report which database object were not reported on.