pgsql: update parallelism in data load script
Update the database build script to detect how many processors are on
the system to determine how much to parallelize the data load.  This is
also assuming the system generating data is separate from the database
system.

Then parallelize the load per table.  In other words, load only one
table at a time to limit how many processes are started to parallelize
the data load.

The datagen program can still be run manually to control how to load the
database.