fix altered mode
The code didn't properly manage the pthread structure when limiting the
number of threads to be created by 1 per warehouse, ignoring the
variable of the number of terminals to create per warehouse.

Instead allow the "altered" execution to create as many threads per
warehouse as the user desires.  The code will still create at least 1
thread per the warehouse range specified by wmin and wmax.  The user
and then control the total number of threads started by adjusting the
number of threads created per warehouse with the tpw flag.