refactor altered mode
Added -L flag to the driver to limit the max number of threads that can
be started.  Note that this can affect non-altered driving modes but is
intended as a knob for altered driving.  This allow a finer grain of
control with how many threads are started in addition to using the -tpw
(districts per warehouse) and -spread (warehouses to skip.)

Updated the run-workload script to calculate rampup time with this new
thread limiting flag.  This also fixed the wait time calculation for
how long the client should take to start up.  Some output messages have
also been updated.