Refactor client code for libpq
Update the build system so automatically build support for libpq if the
libraries exist.  The end goal is to have a single client binary work
with all connectivity libraries that exist on the system.

Scripts and client binary have been updated to take or use the
additional flag to specify which database management system to connect
to.

Data structures have been combined in unions.  Function pointers added
to redirect database specific connectivity calls to the appropriate
functions.

The remaining supported database connectivity libraries still need to
be refactored.