Improvements on dbt2-pgsql-build-db
The new -f flag can be used to execute vacuumdb in FULL mode. The
default mode is now FREEZE: performing vacuum full requires a
huge amount of time and is not really necessary.

vacuumdb can now be executed with multiple jobs. The number of jobs
is automatically calculated, depending on Postgres version and the
number of CPUs.