Add multi-process event driven client2
This experimental client forks a process per detected processor, each
with one database connection, pins each process to a processor and uses
libev to manage connections from the driver.

A unix domain socket is implemented as a queue for incoming driver
connections, but the driver connection is not shared between child
processes once it is pulled from the unix domain socket.