Update perf collection and reporting
Collect only 10 seconds of data, instead of the entire test duration.
Anything longer can be a too much data to process.  Hopefully
immediately after rampup is a good time to collect a small
representative sample.  Maybe the duration needs to be configurable?

Update collection to happen on each node, as we may want to profile the
driver and client in addition to the database system.

The links to the perf reports also has been added back to the reporting
scripts.