Add -F flag to run-workload for driver3 connections per processor
The -F flag will tell driver3 how many database connections to open per
processor.

Update the related Docker files so that the driver container is built
with the necessary libraries so that driver3 can be used against
PostgreSQL.

Also add a new Docker helper script, run3, as an example on how to
execute run-workload with driver3.