Fix driver3 ramp up and stop time calculations
This code was adapted from the original pthread code that was designed
to sleep between each connection opened to the client and each
connection opened to the database where connection rates needed to be
throttled.

This patch simplifies the code further because now only connections
opened to the database need to be throttled, since the client tier is
not used with driver3, and we only open 1 connection per process forked.

Some additional bugs are being squashed at the same time.

* Fix start_time calculation to factor in now many processes are going
  to be started, as opposed to how many warehouses there are.
* Stop sleeping between warehouse and district terminal assignments
  because these are now only memory allocations for libev, as opposed to
  pthread creation and establishing client socket connections.  This was
  also throwing off the stop time calculations because it wasn't
  correctly factored in
* The sleep was being incorrectly re-calculated from being brought in
  from the multi-threaded model to the multi-processor event-driven
  model.  Recalculating it is not necessary at this point.
* The estimated ramp up time in the run-workload script was off by 1.

A couple cosmetic updates:

* Remove unused continue statement.
* Fixed a comment typo.