Rewrite post-process in Julia
Explore using a Julia script for post processing because Julia and the
post-process script package dependencies are license compatible and thus
can be distributed with DBT-2 in an AppImage.

Rename the previous R version with a .r extension, where the Julia
version has a .jl extension.  Use a wrapper script to determine which
language is available on the system, to determine which version of the
script to run.

Also removed an extra space in one out of lines of output from the
original version.