Convert documentation to rst
Convert the documents into reStructuredText.  Then HTML, pdfs and other
formats can be generated from that.

This also creates a single documents, as well as leaves the individual
files to be viewed individually.

A separate Makefile should be easier to maintain than using CMake, since
it's not dependent on any code.

The documentation is still stale and needs updating.