Fix datagen partitioning correctness
Previously the data generator was manually checked and it only
partitioned database deterministically if the scale factor matched the
number of partitions.

The issue doesn't break the workload, but now we are using the pseudo
random data generator correctly to generate deterministic data in
parallel.

Also added some tests and noted them in the documentation.