pl/pgsql default for pgsql stored functions
This is an effort to simplify kit usage.  For the less experienced, it
will be easier to use pl/pgsql stored functions than to use the C
stored functions, which required building the C shared objects on the
database server.  C stored functions are not possible to use on most
DBaaS options either.

The C stored functions are still expected to perform significantly
better and it is suggested that the experienced person looking for the
extra performance use them if possible.