Flip usage of -u flag to specify execution as a privileged user
This is for ease of using the kit by making the default such that the
user may not have privileges to restart, create, or drop a database.  It
is meant to be easier because there is potentially more work to give the
user privileges elevated privileges to perform these tasks before they
can try to run the test in that manner.

Thus advanced usage of the kit will need the use of more flag, as
opposed to having more flags to make the kit easier to use.