AppImage scripting update
Use Z flag for mounting the host source directory for shared volume
labeling.  When it's not needed, it seems ok to have.  Otherwise some
system may experience permission failures when trying to use a
container to build the AppImage.

Also be explicit about the system architecture by matching the container
system.