post-process report p_value of terminal id
Use Chi-Square Goodness of Fit Test to determine if requests from the
driver (i.e. emulated terminals) are coming from warehouse and district
id's in a uniform distribution by displaying their respective p-values.

While 0.05 is a commonly used threshold, it is still left up to the
tester to determine the p-value is sufficient above 0.05 to feel
confident that all warehouses and districts are being used.

The R script was previously just displaying the counts per warehouse and
district.  The Julia script was previously not displaying anything when
VERBOSE was set to to 1.

This is hopefully more useful in validating that the altered mode test
execution is working as expected with the event-driven multi-process
driver.