Remove julia dbt2-post-process.jl
In anticipation of adding SQLite to do post-processing, which can be
used to do Type 1 quantile calculations [1], where Julia's Distribution
package cannot currently do Type 1.  While Julia can be used to do a
Type 1 calculation, in a similar way that SQLite can be use, there is
also the additional challenge of having the scripts also set up required
packages as opposed to relying on the user to prepare the environment
manually.

Ultimately it seems easier to maintain the kit by removing the Julia
post processing script at this time.

[1] https://en.wikipedia.org/wiki/Quantile