Add dbt2 script
The "dbt2" script is for simulating a single "dbt2" command for all
scripts and binaries provided by the kit.  The hope is that this will
make the kit usable with an AppImage (installed as "dbt2") or a direct
installation from source or other packaging mechanisms.

Also merged dbt2-version functionality in to "dbt2" script.