Add ssh to AppImage
Because of the way LD_LIBRARY_CONFIG is used in the AppImage and that
openssl is built for the AppImage, there are potential conflicts with
ssh on the host system.  It seems like the simple path forward is to add
ssh to the AppImage that is built against the version of openssl in the
AppImage.