psql: check if database exists before creating
If a database exists and the user doesn't have permissions, the SQLSTATE
for permissions gets thrown before the existing database code.  To be
able to distinguish between the to, check if the database already exists
before trying to create it.  Then any error from trying to create the
database should be because of a lack of permission.