driver3: reopen database connections on disconnect
The kit will not attempt to reconnecting to the database if a connection is
dropped for the expected duration of the test.

Introduce RECONNECT error code so db.c knows when to attempt reconnecting to
the database.

Add -retry-delay to driver3 to control the wait in milliseconds between
retrying database connections.

Currently only libpq throws RECONNECT error codes.