Specify rows to commit in `build` script
Support specifying the number of rows to commit in a load with the helper
`build`script, so that it passes the requested functionality through to datagen
when loading data.

Further work still needed in some of the database management system scripts to
actually do this in datagen.  This should work for any of the database systems
that look like PostgreSQL, but may not work yet in the rest.