Revert linuxdeploy and update use of appimagetool
One of the linuxdeploy alpha releases might have an interesting elf
compatibility issue with Debian 10, but unfortunately needed something
to work before chasing that.

As part of reverting that, cleaned up and made Container.appimage and
CMakeLists.txt rules a little more intelligent with less redundant
copying of files move some of the commands from Containerfile.appimage
into the CMakeLists.txt so that an AppImage can be built with the same
commands as it is done in a container.