AppImage: include pgsql c user defined funcs
The PostgreSQL C stored functions are packaged in the AppImage, but they
still need to be manually extracted, compiled and installed in order to
before they can be used.

This is intended just to provide everything one might want to use in an
AppImage without having to retrieve anything more from the source
repository.