mysql payment stored procedure: use of of FOR UPDATE and missing c_data
The July 12 didn't look accurate.  It looked like the query is using
c_d_id and c_w_id when searching for customers by last name.

Added FOR UPDATE in a couple places that should help.

Didn't see a reason to not add c_data to the UPDATE statement, per spec.